requirements_win.txt
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.6
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.0
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.4
Markdown==3.1
MarkupSafe==1.1.1
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.25.0
Pillow==6.0.0
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2017.2
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.24.1
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.1.6
xlwt==1.3.0
